+++ CHANGELOG.md
@@ -8,7 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Changelog.md to track changes on this project
+* Changelog.md to track changes on this project
+* Add support for DAW Map Preset
+  * jog wheel is used to change the position of left&right loop locators (press the jog wheel to change the active loop locator)  
+  * Left & Right arrow buttons arround the jog wheel are use to change the tempo
+  * Master fader is used to change the Master Section level (This control surface need to be locked to the Master Section to work properlly, see [Reason doc](http://docs.propellerheads.se/reason10/wwhelp/wwhimpl/js/html/wwhelp.htm#context=EngOpManProjectPro&topic=PreferencesControlSurface))
 
 ### Changed